services/transcription/requirements.txt
# FastAPI and server
fastapi>=0.109.0,<0.115.0
uvicorn[standard]>=0.27.0,<0.32.0

# File handling
python-multipart>=0.0.6
aiofiles>=23.2.1

# Validation
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Speech-to-text (более гибкие версии)
faster-whisper>=1.0.0,<2.0.0

# Speaker diarization (используем последние совместимые версии)
pyannote.audio>=3.1.0,<4.0.0

# Audio processing (используем более новые но CPU-совместимые версии)
torch>=2.1.0,<2.5.0
torchaudio>=2.1.0,<2.5.0

# Utilities
numpy>=1.24.0,<2.0.0
scipy>=1.10.0,<1.15.0
